requirements.txt
flask
flask-restful
gunicorn
RPi.GPIO
